+++ en/events.js
@@ -1,3 +1,38 @@
 export default {
-  'events.settings.title': 'Event settings'
+  'events.upcoming': 'Upcoming events',
+  'events.settings.title': 'Event settings',
+  'events.attendance.attending': 'Attending',
+  'events.attendance.tentative': 'Tentative',
+  'events.attendance.notAttending': 'Not attending',
+  'events.details.name': 'Event name',
+  'events.details.image': 'Event image',
+  'events.details.description': 'Event description / Agenda',
+  'events.details.url': 'Link to more details',
+  'events.details.virtualLink':
+    'Link to virtual location (ie. Zoom, Google Meets, etc.)',
+  'events.details.address': 'Venue address',
+  'events.details.startsAt': 'Event start',
+  'events.details.endsAt': 'Event ends',
+  'events.details.type': 'Event type',
+  'events.type.inPerson': 'In-person',
+  'events.type.virtual': 'Virtual',
+  'events.type.hybrid': 'Hybrid',
+  'events.join.startsIn': 'Event starts {startsAt}', // The startsAt is going to be in a relative time format. Eg. `in 1 minute`, `1 day ago`, etc...
+  'events.join.virtual': 'Join event',
+  'events.selection.origin': 'Event origin',
+  'events.origin.universe': 'Universe',
+  'events.origin.author': 'Author',
+  'events.origin.groups': 'Groups',
+  'events.origin.organizations': 'Organizations',
+  'events.origin.lu': 'Literary Universe events',
+  'events.visibility.public': 'Public - anyone can see this event',
+  'events.visibility.owner': `Owner - only {type, select,
+    group {members of your group}
+    universe {collaborators and fans of the universe}
+    organization {members of your organization}
+    other {you and the people you invite}
+  } can see this event`,
+  'events.visibility.private':
+    'Private - only people you invite will see the event',
+  'events.inviting': 'Can people who see this event invite others?'
 }